% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overrideView.R
\name{get_actiontype}
\alias{get_actiontype}
\title{Create MiniUI for Selection}
\usage{
get_actiontype(a)
}
\arguments{
\item{a}{Dateaframe to act on}
}
\value{
A choice of action
}
\description{
Create MiniUI for Selection
}
