% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overrideView.R
\name{View1}
\alias{View1}
\title{Recreate the View Function}
\usage{
View1(x, title = "")
}
\arguments{
\item{x}{default x View Parameter}

\item{title}{default title View Parameter}
}
\value{
View data object
}
\description{
Recreate the View Function
}
