% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancedView-package.R
\docType{package}
\name{enhancedView-package}
\alias{enhancedView}
\alias{enhancedView-package}
\title{enhancedView: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Tim-Lukas-H/enhancedView}
  \item Report bugs at \url{https://github.com/Tim-Lukas-H/enhancedView/issues}
}

}
\author{
\strong{Maintainer}: Tim Holzapfel \email{TimHolzapfel@outlook.com}

}
\keyword{internal}
