% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View.R
\name{View}
\alias{View}
\title{Enhanced Data Viewer}
\usage{
View(
  file,
  pageLength_default = getOption("enhancedView.pageLength", default = 200),
  default_view = getOption("enhancedView.default_view", default = FALSE),
  theme = getOption("enhancedView.theme", default = "flatly")
)
}
\arguments{
\item{file}{File to be viewed}

\item{pageLength_default}{default page length.}

\item{default_view}{Request the default RStudio data viewer.}

\item{theme}{Drop variables that are not relevant in most situations}
}
\description{
One of the biggest problems of the built-in data viewer is its
limited capability to expand the columns.
}
