% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View.R
\name{View}
\alias{View}
\title{Enhanced Data Viewer}
\usage{
View(file, dev_mode = TRUE)
}
\arguments{
\item{file}{File to be viewed}
}
\description{
One of the biggest problems of the built-in data viewer is its
limited capability to expand the columns.
}
